+++ pyproject.toml
@@ -1,10 +1,10 @@
 [metadata]
-name = "HCPInterface"
+name = "NGPIris"
 [build-system]
 requires = ["setuptools>=42","wheel"]
 build-backend = "setuptools.build_meta"
 [project]
-name = "HCPInterface"
+name = "NGPIris"
 authors = [{name = "Isak Sylvin", email = "isak.sylvin@gu.se"},]
 readme="README.md"
 dynamic = ["version","description"]
