+++ setup.cfg
@@ -1,13 +1,13 @@
 [metadata]
-name=HCPInterface
-version=2.3.3
+name=NGPIris
+version=2.3.4
 long_description=file: README.md
 long_description_content_type=text/markdown
-url=https://github.com/genomic-medicine-sweden/HCPInterface
+url=https://github.com/genomic-medicine-sweden/NGPIris
 author=Isak Sylvin
 author_email=isak.sylvin@gu.se
 project_urls=
-    Bug Tracker = https://github.com/genomic-medicine-sweden/HCPInterface/issues
+    Bug Tracker = https://github.com/genomic-medicine-sweden/NGPIris/issues
 classifiers=
     Programming Language :: Python :: 3
     License :: OSI Approved :: GNU General Public License v2 (GPLv2)
